+++ src/core/model/shapes/layer.cpp
@@ -170,3 +170,23 @@ QIcon model::Layer::static_tree_icon()
 {
     return QIcon::fromTheme("folder");
 }
+
+std::unique_ptr<model::ShapeElement> model::Layer::to_path() const
+{
+    auto clone = std::make_unique<model::Layer>(document());
+
+    for ( BaseProperty* prop : properties() )
+    {
+        if ( prop != &shapes )
+            clone->get_property(prop->name())->assign_from(prop);
+    }
+
+    for ( const auto& shape : shapes )
+    {
+        clone->shapes.insert(shape->to_path());
+        if ( shape->is_instance<model::Modifier>() )
+            break;
+    }
+
+    return clone;
+}